pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django-allauth>=65.8.0",
    "django>=5.2.1",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "requests>=2.32.3",
    "pyjwt>=2.10.1",
    "cryptography>=44.0.3",
    "oauthlib>=3.2.2",
    "urllib3>=2.4.0",
    "flask-login>=0.6.3",
]
